average n fit method implemented
